moving lift methods
