+++ trunk/SUAVE/Analyses/Aerodynamics/AERODAS.py
@@ -14,6 +14,7 @@ from SUAVE.Core import Data, Units
 from SUAVE.Analyses import Process
 from SUAVE.Analyses.Aerodynamics.Process_Geometry import Process_Geometry
 from SUAVE.Methods.Aerodynamics import AERODAS as Methods
+from SUAVE.Methods.Aerodynamics.Common import Higher_Fidelity as Common
 
 # ----------------------------------------------------------------------
 #  AERODAS
@@ -75,7 +76,7 @@ class AERODAS(Markup):
         compute.lift_drag_total                        = Methods.AERODAS_setup.lift_drag_total
         
         compute.lift = Process()
-        compute.lift.total                             = Methods.AERODAS_setup.lift_total
+        compute.lift.total                             = Common.Lift.lift_total
         compute.drag = Process()
         compute.drag.total                             = Methods.AERODAS_setup.drag_total
