+++ trunk/SUAVE/Analyses/Aerodynamics/SU2_Euler.py
@@ -23,6 +23,7 @@ from Results import Results
 
 # The aero methods
 from SUAVE.Methods.Aerodynamics.Common import Fidelity_Zero as Common
+from SUAVE.Methods.Aerodynamics.Common import Higher_Fidelity as Higher_Fidelity
 from Process_Geometry import Process_Geometry
 from SUAVE.Analyses.Aerodynamics.SU2_inviscid import SU2_inviscid
 
@@ -80,8 +81,8 @@ class SU2_Euler(Markup):
         compute.lift = Process()
 
         # Run SU2 to determine lift
-        compute.lift.inviscid                         = SU2_inviscid()
-        compute.lift.total                            = SUAVE.Methods.Aerodynamics.AERODAS.AERODAS_setup.lift_total
+        compute.lift.inviscid                      = SU2_inviscid()
+        compute.lift.total                         = Higher_Fidelity.Lift.lift_total
         
         # Do a traditional drag buildup
         compute.drag = Process()
@@ -100,10 +101,10 @@ class SU2_Euler(Markup):
         compute.drag.compressibility.wings.wing    = Common.Drag.compressibility_drag_wing
         compute.drag.compressibility.total         = Common.Drag.compressibility_drag_wing_total        
         compute.drag.miscellaneous                 = Common.Drag.miscellaneous_drag_aircraft_ESDU
-        compute.drag.untrimmed                     = SUAVE.Methods.Aerodynamics.SU2_Euler.untrimmed
+        compute.drag.untrimmed                     = Higher_Fidelity.Drag.untrimmed
         compute.drag.trim                          = Common.Drag.trim
         compute.drag.spoiler                       = Common.Drag.spoiler_drag
-        compute.drag.total                         = SUAVE.Methods.Aerodynamics.SU2_Euler.total_aircraft_drag
+        compute.drag.total                         = Higher_Fidelity.Drag.total_aircraft_drag
         
         
     def initialize(self):
