+++ trunk/SUAVE/Methods/Aerodynamics/Common/Higher_Fidelity/Drag/__init__.py
@@ -1,0 +1,6 @@
+## @defgroup Methods-Aerodynamics-Common-Higher_Fidelity-Drag Drag
+# Aerodynamic methods contain the functions for the aerodynamic analyses.
+# @ingroup Methods-Aerodynamics-Common-Higher_Fidelity
+
+import total_aircraft_drag
+import untrimmed
