+++ trunk/SUAVE/Methods/Aerodynamics/Common/Higher_Fidelity/Lift/__init__.py
@@ -2,5 +2,4 @@
 # Aerodynamic methods contain the functions for the aerodynamic analyses.
 # @ingroup Methods-Aerodynamics-Common-Fidelity_Zero
 
-from aircraft_total import aircraft_total
-from fuselage_correction import fuselage_correction
+from lift_total import lift_total
