+++ trunk/SUAVE/Methods/Aerodynamics/Common/Higher_Fidelity/__init__.py
@@ -1,0 +1,6 @@
+## @defgroup Methods-Aerodynamics-Common-Higher_Fidelity Higher_Fidelity
+# Aerodynamic methods contain the functions for the aerodynamic analyses.
+# @ingroup Methods-Aerodynamics-Common
+
+import Drag
+import Lift
