+++ trunk/SUAVE/Methods/Aerodynamics/Common/__init__.py
@@ -3,3 +3,4 @@
 # @ingroup Methods-Aerodynamics
 
 import Fidelity_Zero
+import Higher_Fidelity
