+++ trunk/SUAVE/Methods/Aerodynamics/Fidelity_Zero/Lift/__init__.py
@@ -7,5 +7,3 @@ from compute_max_lift_coeff import compute_max_lift_coeff
 from compute_flap_lift import compute_flap_lift
 from compute_slat_lift import compute_slat_lift
 from wing_compressibility_correction import wing_compressibility_correction
-from fuselage_correction import fuselage_correction
-from aircraft_total import aircraft_total
