--- trunk/SUAVE/Methods/Aerodynamics/SU2_Euler/__init__.py
@@ -1,6 +1,0 @@
-## @defgroup Methods-Aerodynamics-SU2_Euler SU2_Euler
-# SU2_Euler contains the function needed to run the SU2 Euler Analysis.
-# @ingroup Methods-Aerodynamics
-
-from total_aircraft_drag import total_aircraft_drag
-from untrimmed import untrimmed
