+++ trunk/SUAVE/Methods/Aerodynamics/Supersonic_Zero/Lift/__init__.py
@@ -1,8 +1,6 @@
 ## @defgroup Methods-Aerodynamics-Supersonic_Zero-Lift Lift
 # Functions to perform low-fidelity lift calculations for supersonics
 # @ingroup Methods-Aerodynamics-Supersonic_Zero
-from vortex_lift import vortex_lift
 
+from vortex_lift import vortex_lift
 from wing_compressibility import wing_compressibility
-from fuselage_correction import fuselage_correction
-from aircraft_total import aircraft_total
