+++ trunk/SUAVE/Methods/Aerodynamics/__init__.py
@@ -6,4 +6,4 @@ import AVL
 import AERODAS
 import Fidelity_Zero
 import Supersonic_Zero
-import SU2_Euler
+import Common
